+++ ui/vite.config.ts
@@ -1,8 +1,12 @@
-import { defineConfig } from 'vite'
+import { defineConfig } from 'vitest/config'
 import react from '@vitejs/plugin-react'
 
 // https://vite.dev/config/
 export default defineConfig({
-  plugins: [react()],
-  publicDir: 'assets'
+  // eslint-disable-next-line @typescript-eslint/no-explicit-any
+  plugins: [react() as any],
+  publicDir: 'assets',
+  test: {
+    environment: 'jsdom'
+  }
 })